+++ packages/react-g/src/reconciler.ts
@@ -1,6 +1,6 @@
 import type { Canvas, Element } from '@antv/g';
 import React from 'react';
-import type { OpaqueHandle } from 'react-reconciler';
+import type { OpaqueHandle, OpaqueRoot } from 'react-reconciler';
 import ReactReconciler from 'react-reconciler';
 import { unstable_now as now } from 'scheduler';
 import { bindShapeEvent, updateProps } from './processProps';
@@ -20,7 +20,7 @@ import type {
   UpdatePayload,
 } from './types';
 
-export const reconcilor = ReactReconciler<
+export const reconciler = ReactReconciler<
   Type,
   Props,
   Container,
@@ -346,7 +346,7 @@ export const reconcilor = ReactReconciler<
   ): void {},
 });
 
-reconcilor.injectIntoDevTools({
+reconciler.injectIntoDevTools({
   // findFiberByHostInstance: () => {},
   // @ts-ignore
   bundleType: process.env.NODE_ENV !== 'production' ? 1 : 0,
@@ -358,6 +358,8 @@ reconcilor.injectIntoDevTools({
     },
   },
 });
+
+const TargetContainerWeakMap = new WeakMap<Element | Canvas, OpaqueRoot>();
 
 /**
  * render react-g component to target g element
@@ -372,7 +374,9 @@ export const render = (
   target: Element | Canvas,
   callback?: (() => void) | null,
 ) => {
-  // @ts-ignore
-  const container = reconcilor.createContainer(target as any, 1, false, null);
-  reconcilor.updateContainer(component, container, null, callback);
+  const container =
+    TargetContainerWeakMap.get(target) ||
+    reconciler.createContainer(target as any, 1, false, null);
+  TargetContainerWeakMap.set(target, container);
+  reconciler.updateContainer(component, container, null, callback);
 };